Finished ChooseSubtree, Split and Insertion.

+++ Leaf.java
@@ -6,10 +6,10 @@
 public class Leaf extends NodeEntry
 {
     private int dataBlockID; //Block index that points to a specific block in dataFile.
-    private int recordID; //Record index that points into a Record in a specific block.
+    private long recordID; //Record index that points into a Record in a specific block.
 
 
-    Leaf(int dataBlockID,int recordID, MBR mbr) {
+    Leaf(int dataBlockID,long recordID, MBR mbr) {
         super(mbr);
         this.dataBlockID = dataBlockID;
         this.recordID = recordID;
@@ -28,7 +28,7 @@ public class Leaf extends NodeEntry
      * Getter function that return the record index in the dataBlock saved in dataFile.
      * @return integer index of a record in current block stored in DataFile.
      */
-    int getRecordID()
+    long getRecordID()
     {
         return recordID;
     }
